integrated new note feature
