+++ packages/frontend/src/pages/private/NotesPage.tsx
@@ -1,33 +1,41 @@
-import React from 'react';
+import React, { useState } from 'react';
 import Page from '../../components/common/layout/Page';
-import { Button, Text, Grid, Card } from '@nextui-org/react';
-import { PlusIcon } from '@heroicons/react/outline';
+import { useHouse } from '../../hooks/useHouse';
+import NoteCardController from '../../components/notes/note/noteCardController';
 import UnderlinedText from '../../components/dashboard/GradientUnderlinedText';
+import { Button } from '@nextui-org/react';
+import { PlusIcon } from '@heroicons/react/outline';
+import NewNoteModal from '../../components/notes/note/newNoteModal';
 
 interface NotesProps {}
 
 const NotesPage: React.FC<NotesProps> = () => {
+  const { name } = useHouse();
+
+  const [createNoteVisible, setCreateNoteVisible] = useState(false);
+  const createNoteHandler = () => setCreateNoteVisible(true);
+
   return (
     <Page>
       <div className="flex justify-between items-center pb-1">
         <UnderlinedText colorClasses="from-gray-800 via-teal-700 to-teal-500 ">
-          <div className="text-lg font-large">{'Notes'}</div>
+          <div className="text-lg font-medium">
+            Create Note
+            <span className="text-teal-500 font-semibold">{name}</span>
+          </div>
         </UnderlinedText>
-        <Button icon={<PlusIcon />}>Create Note</Button>
+        <Button
+          auto
+          size="sm"
+          onClick={createNoteHandler}
+          icon={<PlusIcon className="w-6 h-6" />}
+        ></Button>
       </div>
-      <Grid.Container gap={2} justify="center">
-        <Grid xs={12}>
-          <Card color="primary">
-            <Text h4 color="white" size={30}>
-              Note Name
-            </Text>
-            <Text color="white" size={16}>
-              {' '}
-              Note Type
-            </Text>
-          </Card>
-        </Grid>
-      </Grid.Container>
+      <NewNoteModal
+        createNoteVisible={createNoteVisible}
+        setCreateNoteVisible={setCreateNoteVisible}
+      />
+      <NoteCardController />
     </Page>
   );
 };
